% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webvrr.R
\name{addBackground}
\alias{addBackground}
\title{Add a background to a WebVR widget.
For drawing a skybox sphere, use \code{\link{addInnerTexturedSphere}}}
\usage{
addBackground(vr, backgroundType = c("color", "texture", "cube"),
  backgroundColor = NULL, texture = NULL, shouldRepeat = FALSE,
  repeatX = NULL, repeatY = NULL, cube = list())
}
\arguments{
\item{vr}{A webvr widget. See \code{\link{webvrr}}}

\item{backgroundType}{one of "color", "texture", or "cube"}

\item{backgroundColor}{Hex or numeric color representation. Required if backgroundType is "color".
The function call tolerates string and hexmode representations.}

\item{texture}{a string representing the texture in a base64 data URI or a URL. Required if
backgroundType is "texture".}

\item{shouldRepeat}{should the texture repeat?}

\item{repeatX}{repeat density over X portion of mesh}

\item{repeatY}{repeat density over Y portion of mesh}

\item{cube}{a list of 6 strings representing base64 data URIs or URLs corresponding to a cube map.
required if backgroundType is "cube".}
}
\description{
Add a background to a WebVR widget.
For drawing a skybox sphere, use \code{\link{addInnerTexturedSphere}}
}
