% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webvrr.R
\name{addInnerTexturedSphere}
\alias{addInnerTexturedSphere}
\title{Puts the viewer in a big sphere with a texture painted on the inside.
See https://threejs.org/docs/#api/geometries/SphereGeometry}
\usage{
addInnerTexturedSphere(vr, texture = NULL, `repeat` = list(repeatX = 32,
  repeatY = 32), radius = 5, widthSegments = 32, heightSegments = 32)
}
\arguments{
\item{vr}{a webvrr widget. See \code{\link{webvrr}}}

\item{texture}{a string representing a texture in base64 data URI format or a URL.
if not provided, a wireframe will be used}

\item{repeat}{If false, do not repeat. Otherwise, should be a list in the following
format: \code{list( repeatX = 32, repeatY = 32)}. Replace 32 with your
preferred repeats.}

\item{radius}{sphere radius.}

\item{widthSegments}{how many width segments to draw}

\item{heightSegments}{how many height segments to draw}
}
\description{
Puts the viewer in a big sphere with a texture painted on the inside.
See https://threejs.org/docs/#api/geometries/SphereGeometry
}
