% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webvrr.R
\name{addToScene}
\alias{addToScene}
\title{Add an arbitrary object to the scene
Internal.}
\usage{
addToScene(vr, call = list())
}
\arguments{
\item{vr}{a webvrr widget}

\item{call}{a specially crafted call object}
}
\description{
Add an arbitrary object to the scene
Internal.
}
